Move methods to merchantrepository
